test(sync-core): pin sync()'s canister-call pattern as a regression test
The benches in bench_sync.rs were `#[ignore]`d and asserted nothing beyond
"sync() returned Ok". They're fast (~1s even for the 50 MB case) and the
canister-call pattern they measure — per-method call counts + total Candid
arg bytes — is fully deterministic (fixed file sizes, fixed LCG seed,
deterministic chunking + Candid encoding), independent of build profile and
host.

Un-ignore them so they run in the normal suite, and assert that fingerprint
against a committed golden. A diff now flags a real change in the calls
sync() emits — e.g. lost chunk-batching (inflated upload_chunks) or lost
operation-grouping (inflated execute_operations). Wall-clock stays
informational (printed under --nocapture), never asserted.

Rename bench_sync.rs -> call_pattern.rs and drop the now-misleading bench_
prefix from the tests, helper, and mock.
